fix(mongodb): resolve clippy -D warnings in collection import/export
module-level result_large_err allow, manual_range_contains, redundant
closures, collapsible if, is_multiple_of, needless into_iter, type
alias for json preview output, test lint fixes, and a redundant usize
cast in mongo_driver bulk write error mapping

+++ crates/dbx-core/src/db/mongo_driver.rs
@@ -2574,7 +2574,7 @@ fn insert_write_errors(error: &mongodb::error::Error) -> Vec<MongoBulkWriteError
         ErrorKind::BulkWrite(failure) => failure
             .write_errors
             .iter()
-            .map(|(index, error)| write_error_entry(Some(*index as usize), error.code, &error.message))
+            .map(|(index, error)| write_error_entry(Some(*index), error.code, &error.message))
             .collect(),
         _ => Vec::new(),
     };
